chore: Refactor repository imports and enhance database performance
