+++ src/main/java/com/hinderegger/steaminventorytracker/service/ItemService.java
@@ -5,8 +5,8 @@ import com.hinderegger.steaminventorytracker.model.Price;
 import com.hinderegger.steaminventorytracker.model.PriceTrend;
 import com.hinderegger.steaminventorytracker.persistence.ItemEntity;
 import com.hinderegger.steaminventorytracker.persistence.PriceEntity;
-import com.hinderegger.steaminventorytracker.repository.jpa.ItemRepository;
-import com.hinderegger.steaminventorytracker.repository.jpa.PriceRepository;
+import com.hinderegger.steaminventorytracker.repository.ItemRepository;
+import com.hinderegger.steaminventorytracker.repository.PriceRepository;
 import com.hinderegger.steaminventorytracker.service.mapper.ItemMapper;
 import java.time.temporal.ChronoUnit;
 import java.util.List;
