+++ src/test/java/com/hinderegger/steaminventorytracker/RepositoryTest.java
@@ -8,8 +8,8 @@ import com.hinderegger.steaminventorytracker.configuration.SteamConfiguration;
 import com.hinderegger.steaminventorytracker.model.Item;
 import com.hinderegger.steaminventorytracker.persistence.ItemEntity;
 import com.hinderegger.steaminventorytracker.persistence.PriceEntity;
-import com.hinderegger.steaminventorytracker.repository.jpa.ItemRepository;
-import com.hinderegger.steaminventorytracker.repository.jpa.PriceRepository;
+import com.hinderegger.steaminventorytracker.repository.ItemRepository;
+import com.hinderegger.steaminventorytracker.repository.PriceRepository;
 import com.hinderegger.steaminventorytracker.service.*;
 import com.hinderegger.steaminventorytracker.service.mapper.ItemMapper;
 import com.hinderegger.steaminventorytracker.service.strategy.AsyncSteamRequestStrategy;
