+++ src/test/java/com/hinderegger/steaminventorytracker/service/BuyInfoServiceTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import com.hinderegger.steaminventorytracker.model.BuyInfo;
 import java.util.List;
 
-import com.hinderegger.steaminventorytracker.repository.jpa.BuyInfoRepository;
+import com.hinderegger.steaminventorytracker.repository.BuyInfoRepository;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
